+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## v1.0.1
 
-- Bumping dependencies
+- Bump dependencies
 - Migrate to using `getColorByName` instead of `getColorForTheme` to support Grafan 9
 - Migrate to using `GrafanaThema2` instead of `GrafanaTheme`
-- Bumped minimum supported Grafana version to `8.+`
+- Migrate to using `convertOldAngularValueMappings` from `@grafana/data` package instead of `convertOldAngularValueMapping` from `@grafana/ui` package
+- Bump minimum supported Grafana version to `8.+`
+- Remove dependency for `@grafana/ui` as not needed any more
 
 ## v1.0.0